🗑 deprecated _lastProcessedRevlogId
rather than boosting cards since last boosted reviews,
boost cards reviewed since last week.

This is a stability fix. we tracked what cards to boost with '_lastProcessedRevlogId'. It gets updated on each batch boost
But it's quite fuzzy when to update that variable and when it is safe to trust it.

So we deprecated that.